+++ Package.swift
@@ -6,6 +6,7 @@ let package = Package(
     name: "HexToColor",
     platforms: [
         .iOS(.v12),
+        .macOS(.v10_13),
     ],
     products: [
         .library(
@@ -17,12 +18,14 @@ let package = Package(
         .target(
             name: "HexToColor",
             path: "HexToColor",
+            exclude: ["Info.plist"],
             sources: ["Hex.swift"]
         ),
         .testTarget(
             name: "HexToColorTests",
             dependencies: ["HexToColor"],
             path: "HexToColorTests",
+            exclude: ["Info.plist"],
             sources: ["HexToColorTests.swift"]
         ),
     ],